Leverage express-state's `cache` option to improve expose() performance
This updates how Express YUI exposes data to the client by using app-
scoped exposing and caching the result to increase the performance of
serialization.

A non-enumerable `__modified__` flag has been added to the YUI `config`
object which is used by the methods that modify the `config` to signal
that it seems to be re-exposed and re-serialized by express-state.